pyproject.toml
[tool.poetry]
name = "fibonacci"
version = "0.1.0"
description = ""
authors = ["Nikita Antonenkov <krigar1184@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.0"
python-dotenv = "^0.9.1"
coverage = "^4.5"
codecov = "^2.0"
redis = "^3.0"
mockredispy = "^2.9"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
ipdb = "^0.11.0"
mockredispy = "^2.9"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
